==> Bridgestone/10-data-loading.ipynb <==
import pandas as pd
import math
from pathlib import Path

pd.set_option("display.max_columns",None)

store = pd.read_csv('data/p_dsi/teams2022/bridgestone_data/data/store.csv', sep='|')
vehicle = pd.read_csv('data/p_dsi/teams2022/bridgestone_data/data/vehicle_sample.csv').drop('Unnamed: 0',1)
product = pd.read_csv('data/p_dsi/teams2022/bridgestone_data/data/product.csv',sep='|')
individual = pd.read_csv('data/p_dsi/teams2022/bridgestone_data/data/individual_sample.csv').drop('Unnamed: 0',1)
individual.rename(columns={'MZB_INDIV_ID': 'INDIV_ID'}, inplace=True)

# get the mega table that joined all datasets in the path
def joiner(store,vehicle,product,individual):
    file_path = 'data/p_dsi/teams2022/bridgestone_data/data/'
    #output_file = 'mega_table.csv' 
    p = Path(file_path)
    list = []
    for sales_name in p.rglob('sales*.csv'): 
        list.append(sales_name)
    all_sales = pd.read_csv(list[0]).drop('Unnamed: 0',1)
    for i in range(1,len(list)):
        all_sales = pd.concat([all_sales,pd.read_csv(list[i]).drop('Unnamed: 0',1)])
    all_sales = pd.merge(all_sales,store,'left',["STORE_ID", "STORE_ID"])
    all_sales = pd.merge(all_sales,vehicle,'left',["VEHICLE_ID", "VEHICLE_ID"])
    all_sales = pd.merge(all_sales,product,'left',["ARTICLE_ID", "ARTICLE_ID"])
    all_sales = pd.merge(all_sales,individual,'left',["INDIV_ID", "INDIV_ID"])
    all_sales.to_csv('mega_df.csv',index=False)
    return(all_sales)

# 4M rows runs in 10 seconds using colab GPU. 
# Sample is ~ 1/150th of the size of actual so expect 1500s (25 minutes) for largest joins 
joiner(store,vehicle,product,individual)

df = pd.read_csv('mega_df.csv')

# filter a new table with start date and end date
def Date_filter(df, startYear, startMonth, endYear, endMonth,file_name):
    df['Year'] = df['DATE'].str.split('-', expand=True)[0]
    df['Month'] = df['DATE'].str.split('-', expand=True)[1]
    if startYear == endYear:
        df_mod = df[(df['Month'].astype(int) >= startMonth) & 
                (df['Month'].astype(int) <= endMonth) & 
                (df['Year'].astype(int) == startYear)]
        #df_mod.to_csv(file_name)
        return(df_mod)
    elif startYear <= endYear:
        df1 = df[(df['Year'].astype(int) >= startYear+1) & 
                 (df['Year'].astype(int) <= endYear-1)]
        df2 = df[(df['Month'].astype(int) >= startMonth) & 
                (df['Month'].astype(int) <= 12) & 
                (df['Year'].astype(int) == startYear)]
        df3 = df[(df['Month'].astype(int) >= 1) & 
                (df['Month'].astype(int) <= endMonth) & 
                (df['Year'].astype(int) == endYear)]
        df_mod = pd.concat([df1, df2, df3])
        df_mod.to_csv(file_name,index=False)
        return(df_mod)

Date_filter(df,2016,2,2018,3,file_name = 'Feb_2016-Mar_2018.csv')

#filter new training set and testing set for the table with start date and end date
from sklearn.model_selection import train_test_split

#data = pd.read_csv('StartMonth_StartYear-EndMonth_Year.csv')
data = pd.read_csv('Feb_2016-Mar_2018.csv')

def split_dataset(data,train_size):
    train_data, test_data = train_test_split(data, train_size = train_size, random_state = 123)
    return train_data, test_data

training,testing = split_dataset(data,0.7)

training.to_csv("train_Feb_2016-Mar_2018.csv",index = False)
testing.to_csv("test_Feb_2016-Mar_2018.csv",index = False)
