==> Bridgestone/22_feature_engineering_full.ipynb <==
import pandas as pd
import numpy as np
import random

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Taking a random sample from the dataset

# the random sample is skipping 999 rows and taking 1 row on average 
dy = pd.read_csv(
         '../../../JanOct18_trimmed.csv')#, skiprows = lambda i: i>0 and random.random() > 0.1)

# dy16 = pd.read_csv(
#          '../../../JanOct16_trimmed.csv', skiprows = lambda i: i>0 and random.random() > 0.001)

# dy17 = pd.read_csv(
#          '../../../JanOct17_trimmed.csv', skiprows = lambda i: i>0 and random.random() > 0.001)

# dy18 = pd.read_csv(
#          '../../../JanOct18_trimmed.csv', skiprows = lambda i: i>0 and random.random() > 0.001)


dn = pd.read_csv(
         '../../../Nov17_trimmed.csv')
#,         skiprows=lambda i: i>0 and random.random() > 0.1)

vehicles = dy.drop_duplicates(subset = 'VEHICLE_ID',keep='first')

targets = dn[dn['PROD_GROUP_CODE'] == 5]['INDIV_ID'].unique()


year_indivs = pd.DataFrame({'indiv':dy['INDIV_ID'].unique()})

year_indivs['label'] = np.where(year_indivs['indiv'].isin(targets), 1, 0) 

year_indivs['cust_in_top_10_sales'] = np.where(year_indivs['indiv'].isin(dy.INDIV_ID.value_counts().nlargest(int(len(year_indivs)/10))),1,0)

year_indivs.label.value_counts()

len(year_indivs)

# def label_training__data(year_file_name, nov_file_name):
#     year_dat = pd.read_csv(year_file_name)
#     nov_dat = nov_file_name

### Feature : is the store among the top performers regarding total number of transactions in general?

top_stores_trans_list = dy.STORE_ID.value_counts().nlargest(int(len(dy)/100)).index

len(top_stores_trans_list)

get individuals visiting those stores

indiv_store_list = dy[dy['STORE_ID'].isin(top_stores_trans_list)]['INDIV_ID']

year_indivs['store_top_tran'] = np.where(year_indivs.indiv.isin(indiv_store_list), 1, 0)

# sanity check
year_indivs['store_top_tran'].value_counts()

### Feature : is the store among the top performers regarding total number of TIRES transactions ?

tires = dy.query('PROD_GROUP_CODE == 5')

top_stores_tires_list = tires.STORE_ID.value_counts().nlargest(int(len(dy)/100)).index

top_stores_tires_list

get individuals visiting those stores

indiv_store_tires_list = dy[dy['STORE_ID'].isin(top_stores_tires_list)]['INDIV_ID']

year_indivs['store_top_tires_tran'] = np.where(year_indivs.indiv.isin(indiv_store_tires_list), 1, 0)

# sanity check
year_indivs['store_top_tires_tran'].value_counts()

year_indivs.head()

indiv_store_tires_list = dy[dy['STORE_ID'].isin(top_stores_tires_list)]['INDIV_ID']

year_indivs['store_top_tires_tran'] = np.where(year_indivs.indiv.isin(indiv_store_tires_list), 1, 0)

# sanity check
year_indivs['store_top_tires_tran'].value_counts()

## Feature: has this vehicle recently had a tire service? 

consider adding a certain time limit 

### list of vehicles with recent tires' services 

vhc_tire_svc_array = dy.query('CATEGORY_DESC == "Tire Services"')['VEHICLE_ID'].unique()

indiv_vhc_svc_list = dy[dy['VEHICLE_ID'].isin(vhc_tire_svc_array)]['INDIV_ID']

year_indivs['vehicle_tire_svc'] = np.where(year_indivs.indiv.isin(indiv_vhc_svc_list), 1, 0)

# sanity check
year_indivs['vehicle_tire_svc'].value_counts()

## Feature: has anyone purchased a tire for this vehicle before?

we need to settle on the definition of tire purchase, is it 5 only or after removing others as well?

vhc_tire_pur_array = dy.query('PROD_GROUP_DESC == "Tires" & SEGMENT_DESC != "Others"')['VEHICLE_ID'].unique()

vhc_tire_pur_array

indiv_vhc_pur_list = dy[dy['VEHICLE_ID'].isin(vhc_tire_pur_array)]['INDIV_ID']

year_indivs['vhc_tire_purchase'] = np.where(year_indivs.indiv.isin(indiv_vhc_pur_list), 1, 0)

# sanity check
year_indivs['vhc_tire_purchase'].value_counts()

## Feature : when was the last time this vehicle bought a new tire? [last year - 2 years - 3 years]

dy['DATE'] = pd.to_datetime(dy['DATE'])

dy['year'] = pd.DatetimeIndex(dy['DATE']).year
dy['month'] = pd.DatetimeIndex(dy['DATE']).month

dy_vhc_tire_purch = dy.query('PROD_GROUP_DESC == "Tires" & SEGMENT_DESC != "Others"')

month_look = dy_vhc_tire_purch.groupby('VEHICLE_ID').agg({'month':['max'] })

Max is the latest month the tire was purchased. If the last month to purchase a tire was before 6, the vehicle may need to replace tires. 
Of course, this needs a change, but after we combine datasets of different years. 

month_look.columns = ['last_mo_tire_purch']

vhc_early_purch_tires = month_look.reset_index().query('last_mo_tire_purch < 6')['VEHICLE_ID']

indiv_vhc_early_pur_list = dy[dy['VEHICLE_ID'].isin(vhc_early_purch_tires)]['INDIV_ID']

year_indivs['vhc_early_tire_purchase'] = np.where(year_indivs.indiv.isin(indiv_vhc_early_pur_list), 1, 0)

# sanity check
year_indivs['vhc_early_tire_purchase'].value_counts()

# Alert 

too many vehicles with ID =1 

### Here I need to understand what to do with units 

## Feature: top customers regarding overall transactions  

top_cust_overall_list = dy.INDIV_ID.value_counts().nlargest(int(len(dy)/10))

year_indivs['top_cust_overall'] = np.where(year_indivs['indiv'].isin(top_cust_overall_list), 1, 0)

## Feature: top customers regarding tires' purchase 

top_cust_tire_purch_array = dy_vhc_tire_purch.INDIV_ID.value_counts().nlargest(int(len(dy_vhc_tire_purch)/10)).index

year_indivs['top_cust_tire_purch'] = np.where(year_indivs['indiv'].isin(top_cust_tire_purch_array), 1, 0)

year_indivs['top_cust_tire_purch'].value_counts()

year_indivs.head(10)

dy.head(10)

Test

top_make_tire_purch_array = dy_vhc_tire_purch.MAKE.value_counts().nlargest(int(len(dy_vhc_tire_purch)/20)).index.tolist()[0:10]

top_make_tire_purch_array

dy['MAKE'] = dy['MAKE'].astype('str')

indiv_make_tire_purch = dy[dy['MAKE'].isin(top_make_tire_purch_array)]['INDIV_ID']

year_indivs['indiv_make_tire_pch'] = np.where(year_indivs.indiv.isin(indiv_make_tire_purch), 1, 0)

year_indivs['indiv_make_tire_pch'].value_counts()

year_indivs.head(10)

dy['tire_purchase'] = np.where((dy['PROD_GROUP_DESC'] == "Tires") & (dy['SEGMENT_DESC'] != "Others"), 1, 0)

df2 = dy.groupby(['MAKE', 'tire_purchase']).agg({'tire_purchase': 'count'})
df3 = df2.groupby(level=0).apply(lambda x:100 * x / float(x.sum()))
df3.columns = ['total'] # I need to rename the column to handle multiindex caused by groupby 

top_make_tires_purch_prop_array = df3.reset_index().query('tire_purchase == 1').sort_values(by = 'total', ascending = False).nlargest(int(len(df3)/10),'total')['MAKE'].tolist()[0:1000]
                                                                                                                                      
indiv_make_tire_purch_prop = dy[dy['MAKE'].isin(top_make_tires_purch_prop_array)]['INDIV_ID']

year_indivs['make_prop_tire'] = np.where(year_indivs.indiv.isin(indiv_make_tire_purch_prop), 1, 0)

year_indivs['make_prop_tire'].value_counts()

df2 = dy.groupby(['MODEL', 'tire_purchase']).agg({'tire_purchase': 'count'})
df3 = df2.groupby(level=0).apply(lambda x:100 * x / float(x.sum()))
df3.columns = ['total'] # I need to rename the column to handle multiindex caused by groupby 

dy['MODEL'] = dy['MODEL'].astype('str')

top_MODEL_tires_purch_prop_array = df3.reset_index().query('tire_purchase == 1').sort_values(by = 'total', ascending = False).nlargest(int(len(df3)/10),'total')['MODEL'].tolist()[0:1000]

indiv_MODEL_tire_purch_prop = dy[dy['MODEL'].isin(top_MODEL_tires_purch_prop_array)]['INDIV_ID']
                                                                                                                                       
year_indivs['MODEL_prop_tire'] = np.where(year_indivs.indiv.isin(indiv_MODEL_tire_purch_prop), 1, 0)
                                                                                                                                       
year_indivs['MODEL_prop_tire'].value_counts()                                                                                                                                    

df2 = dy.groupby(['SUB_MODEL', 'tire_purchase']).agg({'tire_purchase': 'count'})
df3 = df2.groupby(level=0).apply(lambda x:100 * x / float(x.sum()))
df3.columns = ['total'] # I need to rename the column to handle multiindex caused by groupby

dy['SUB_MODEL'] = dy['SUB_MODEL'].astype('str')

top_SUB_MODEL_tires_purch_prop_array = df3.reset_index().query('tire_purchase == 1').sort_values(by = 'total', ascending = False).nlargest(int(len(df3)/10), 'total')['SUB_MODEL'].tolist()[0:1000]            

indiv_SUB_MODEL_tire_purch_prop = dy[dy['SUB_MODEL'].isin(top_SUB_MODEL_tires_purch_prop_array)]['INDIV_ID']
                                                                                                                                               
year_indivs['SUB_MODEL_prop_tire'] = np.where(year_indivs.indiv.isin(indiv_SUB_MODEL_tire_purch_prop), 1, 0)
                                                                                                                                               
year_indivs['SUB_MODEL_prop_tire'].value_counts()

year_indivs.head(10)

year_indivs.to_csv('individuals18.csv')

#
