==> Bridgestone/24_feature_engineering_round3.ipynb <==
import pandas as pd
import numpy as np
import random

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

dy = pd.read_csv(
         '../../../AprOct15_trimmed.csv')

dn = pd.read_csv(
         '../../../Nov15_trimmed.csv')

targets = dn[dn['PROD_GROUP_CODE'] == 5]['INDIV_ID'].unique()
year_indivs = pd.DataFrame({'indiv':dy['INDIV_ID'].unique()})
year_indivs['label'] = np.where(year_indivs['indiv'].isin(targets), 1, 0) 

## Segment
### Example of 2015 trimmed

segment_tire_sales = dy[dy['PROD_GROUP_CODE']==5].groupby(['SEGMENT_DESC'])['SALES'].count().to_frame()

segment_tire_sales.rename(columns = {'SALES': 'NUMBER_OF_TIRE_SALES'},inplace = True)

top_5_segment_tire_sales = segment_tire_sales.sort_values('NUMBER_OF_TIRE_SALES', ascending = False)[:5]

top_10_segment_tire_sales = segment_tire_sales.sort_values('NUMBER_OF_TIRE_SALES', ascending = False)[:10]

#### Top 5

top_5_segment_tire_sales

top_5_segment_tire_sales.reset_index(inplace = True)
top_5_segment_tire_sales = top_5_segment_tire_sales['SEGMENT_DESC']

indiv_top_5_segment_tire_sales = dy[dy['SEGMENT_DESC'].isin(top_5_segment_tire_sales)]['INDIV_ID']

year_indivs['top_5_segment'] = np.where(year_indivs.indiv.isin(indiv_top_5_segment_tire_sales), 1, 0)

year_indivs['top_5_segment'].value_counts()

#### Top 10

top_10_segment_tire_sales

top_10_segment_tire_sales.reset_index(inplace = True)
top_10_segment_tire_sales = top_10_segment_tire_sales['SEGMENT_DESC']

top_10_segment_tire_sales

indiv_top_10_segment_tire_sales = dy[dy['SEGMENT_DESC'].isin(top_10_segment_tire_sales)]['INDIV_ID']

year_indivs['top_10_segment'] = np.where(year_indivs.indiv.isin(indiv_top_10_segment_tire_sales), 1, 0)

year_indivs['top_10_segment'].value_counts()
