==> Bridgestone/26-supp-features.ipynb <==
import pandas as pd 
import numpy as np

path = '../../supp_data/'

%%bash 

cd ../../supp_data

ls

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

# !python -m pip install "dask[complete]" 

## Loading all supp datasets

# acc_miles15 = pd.read_excel(path + 'road_perc_accepted_miles15.xlsx', index_col = 0)
# acc_miles16 = pd.read_excel(path + 'road_perc_accepted_miles16.xlsx', index_col = 0)
acc_miles17 = pd.read_excel(path + 'road_perc_accepted_miles17.xlsx', index_col = 0)
# acc_miles18 = pd.read_excel(path + 'road_perc_accepted_miles18.xlsx', index_col = 0)

# acc_miles = pd.concat([acc_miles15, acc_miles16, acc_miles17, acc_miles18])
acc_miles = acc_miles17

# for the sake of simplicity, I will data from 2017 only for acceptable miles. 2017 acceptable miles can predict both 2017(training) and 2018 (testing) 

weather = pd.read_csv(path + 'weather_by_state.csv', index_col = 0)

miles_covered = pd.read_csv(path + 'daily_miles_by_state_county_tract.csv', index_col = 0)

zip_county_conversion = pd.read_csv(path + 'ZIP-COUNTY-FIPS_2017-06.csv', index_col = 0)

state_to_code = pd.read_csv(path + 'state_to_code.csv', index_col = 0)

acc_miles.head()

From this dataset, we just need the percentage acceptable miles. Our theory is; the higher the percentage, the safer the surface on tires. 

acc_miles.reset_index(inplace = True)

acc_miles.head()

acc_miles = acc_miles[['index', 'Percent Acceptable']]

acc_miles = acc_miles.iloc[2:, :]

acc_miles.head()

state_to_code.head()

acc_miles = acc_miles.merge(state_to_code, left_on = 'index', right_index = True)
acc_miles.head()

acc_miles_codes = acc_miles[['Code', 'Percent Acceptable']]

acc_miles_codes.head()

acc_miles_codes.to_csv(path + 'curated_supp_data/accept_miles.csv')

## Working on weather data

weather.head()

weather.year.value_counts()

For now, I will not need the station granularity. Thus, I will group by State

weather_state = weather.groupby(['state', 'year', 'month']).agg({'avg_temp' : 'mean', 'avg_max_temp' : 'mean',
       'avg_prcp' : 'mean', 'avg_snow_depth' : 'mean'}).reset_index()

weather_state.head()

weather_state.shape

weather_state.to_csv(path + 'curated_supp_data/weather_state.csv')

## MIles Covered 

miles_covered = miles_covered[['est_vmiles']]
miles_covered = miles_covered.dropna().reset_index()
miles_covered.head()

miles_covered['STCOUNTYFP'] = miles_covered.geocode.astype(str).str[:5].astype(int)

miles_covered = miles_covered.groupby('STCOUNTYFP').agg({'est_vmiles': 'mean'}).reset_index()

miles_covered.head()

zip_county_conversion.head()

zip_county_conversion = zip_county_conversion.reset_index()[['ZIP', 'STCOUNTYFP']]

zip_county_conversion.head()

zip_county_conversion = zip_county_conversion.merge(miles_covered, left_on = 'STCOUNTYFP', right_on = 'STCOUNTYFP', how = 'inner')

zip_county_conversion = zip_county_conversion[['ZIP', 'est_vmiles']]

zip_county_conversion.head()

zip_county_conversion.to_csv(path + 'curated_supp_data/zip_county.csv')
