==> Bridgestone/33-modeling-15-16.ipynb <==
## Modeling 2015 + 2016

import pandas as pd

import numpy as np
import random
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score
from sklearn.metrics import recall_score
from sklearn.metrics import precision_score
from sklearn.metrics import roc_auc_score

from sklearn.model_selection import train_test_split, GridSearchCV

data1 = pd.read_csv('../individuals15.csv', index_col = 0)
data2 = pd.read_csv('../individuals16.csv', index_col = 0)

print(data1.shape)
print(data2.shape)

data = pd.concat([data1, data2])
data

test = pd.read_csv('../individuals17.csv', index_col = 0)

# train x = drop the indiv and label (should be 12 columns)
# train y = just the label column
# test x = drop the indiv and label (should be 12 columns) on test
# test y = just the label column on test

x_train = data.drop(columns = ['indiv', 'label'])
y_train = data['label']
x_test = test.drop(columns = ['indiv', 'label'])
y_test = test['label']

model = RandomForestClassifier()
model.fit(x_train,y_train)
yhat = model.predict(x_test)

print("Accuracy Score: ", accuracy_score(y_test, yhat))
print("Precision Score: ", precision_score(y_test, yhat))

# Prediction

df = pd.DataFrame(model.predict_proba(x_test))

df.head()

df_sort = df.sort_values(by = [1],axis = 0,ascending = False)

df_sort[df_sort[1] > 0.0586].shape
