==> Bridgestone/34-modeling-17-16.ipynb <==
import pandas as pd
df17=pd.read_csv('individuals17.csv', index_col = 0)
df17.head()

df16=pd.read_csv('individuals16.csv', index_col = 0)
df16.head()

x_train = df17.drop(columns = ['indiv', 'label'])
y_train = df17['label']
x_test = df16.drop(columns = ['indiv', 'label'])
y_test = df16['label']

from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score
model = RandomForestClassifier()
model.fit(x_train,y_train)
yhat = model.predict(x_test)

print("Accuracy Score: ", accuracy_score(y_test, yhat))

pred = pd.DataFrame(model.predict_proba(x_test))
pred.head()

pred_sort=pred.sort_values(by = [1],axis = 0,ascending = False)
pred_sort.head()

total=pd.concat([df16,pred],axis=1)
total_sort=total.sort_values(by = [1],axis = 0,ascending = False)
total_sort.head()

import numpy as np
top10000=total_sort.head(10000)
print("The proportion of label 1 within the top 10000 observations with highest probability of label 1 is ",np.mean(top10000["label"]==1))
