==> Bridgestone/41-modeling-2015-RF-Adaboost.ipynb <==
## Pkg Loading 

import pandas as pd
import numpy as np
import random

from sklearn.model_selection import train_test_split, GridSearchCV

from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score
from sklearn.metrics import recall_score
from sklearn.metrics import precision_score
from sklearn.metrics import roc_auc_score

from sklearn.metrics import confusion_matrix

from sklearn.ensemble import AdaBoostClassifier

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Data Loading 

dy = pd.read_csv('../../AprOct15_trimmed.csv')

dn = pd.read_csv('../../Nov15_trimmed.csv')

dy.head()

dn.head()

indiv = pd.read_csv('individuals15.csv', index_col = 0)

indiv_test = pd.read_csv('individuals18.csv', index_col = 0)

indiv_test.shape

indiv_test.head()

indiv.head()

indiv.shape

def split_dataset(data,train_size):
    train_data, test_data = train_test_split(data, train_size = train_size, random_state = 123)
    return train_data, test_data

training,testing = split_dataset(indiv,0.7)

training.shape

testing.shape

X_train = training.drop(['indiv', 'label'], axis=1)
y_train = training['label']
X_valid = testing.drop(['indiv', 'label'], axis=1)
y_valid = testing['label']

X_train.shape, y_train.shape, X_valid.shape, y_valid.shape

## Modeling-RF

rf = RandomForestClassifier()
rf.fit(X_train,y_train)

rf_yhat = rf.predict(X_valid)

print("Training Score: ", rf.score(X_train, y_train))

print(confusion_matrix(y_valid,rf_yhat))
print("Accuracy Score: ", accuracy_score(y_valid, rf_yhat))
print("Precision Score: ",precision_score(y_valid, rf_yhat))

## Prediction for test set

df = pd.DataFrame(rf.predict_proba(X_test))

df.head()

df_sort = df.sort_values(by = [1],axis = 0,ascending = False)

df_sort.head()

df_sort[df_sort[1] > 0.0586].shape

df_sort[df_sort[1] > 0.0585].shape

df_sort[(df_sort[1] > 0.0585)&(df_sort[1] < 0.0586)].shape

## Modeling-Adaboost

adb = AdaBoostClassifier()
adb.fit(X_train, y_train) 

adb_yhat = adb.predict(X_valid)

print("Training Score: ", adb.score(X_train, y_train))

print(confusion_matrix(y_valid,adb_yhat))
print("Accuracy Score: ", accuracy_score(y_valid, adb_yhat))
print("Precision Score: ",precision_score(y_valid, adb_yhat))

## Prediction for test set

adb_df = pd.DataFrame(adb.predict_proba(X_test))

adb_df_sort = adb_df.sort_values(by = [1],axis = 0,ascending = False)

adb_df_sort.head()

adb_df_sort[adb_df_sort[1] > 0.48593].shape

adb_df_sort[adb_df_sort[1] > 0.48592].shape

adb_df_sort[(adb_df_sort[1] > 0.48592)&(adb_df_sort[1] < 0.48593)].shape
