==> Bridgestone/60-Generate-List.ipynb <==
import numpy as np
import pandas as pd
import sklearn

from sklearn.ensemble import RandomForestClassifier
from sklearn.datasets import make_classification

dat15 = pd.read_csv('individuals15.csv')
dat16 = pd.read_csv('individuals16.csv')
dat17 = pd.read_csv('individuals17.csv')
dat18 = pd.read_csv('individuals18.csv')

dat15 = dat15.drop('Unnamed: 0',1)
dat16 = dat16.drop('Unnamed: 0',1)
dat17 = dat17.drop('Unnamed: 0',1)
dat18 = dat18.drop('Unnamed: 0',1)

dat_final = pd.concat([dat15,dat16,dat17])

final_model = RandomForestClassifier(max_depth = 10,n_estimators = 100,min_samples_split = 2,min_samples_leaf = 2)

X_train = dat_final.drop(columns = ['indiv','label'])
y_train = dat_final['label']
X_test = dat18.drop(columns = ['indiv'])

indivs = dat18['indiv']
preds = model.predict_proba(X_test)
preds = pd.DataFrame(preds)
preds = preds.join(indivs)
pred100k = preds.sort_values(0).head(100000)['indiv']

pred100k.to_csv('final_list.csv')
